Revert "switcher代码结构优化和部分功能实现"
