--- app/components/icon/closeIcon.tsx
@@ -1,20 +1,0 @@
-const CloseIcon = ({ color = "#000000" }: { color: string }) => {
-  return (
-    <svg
-      viewBox="0 0 1024 1024"
-      version="1.1"
-      xmlns="http://www.w3.org/2000/svg"
-      p-id="4794"
-      width="16"
-      height="16"
-    >
-      <path
-        d="M567.7 509.8l253.2-253.2c12.3-12.3 12.3-32.1 0-44.4a31.325 31.325 0 0 0-44.4 0L523.3 465.4 270.1 212.2a31.325 31.325 0 0 0-44.4 0 31.325 31.325 0 0 0 0 44.4l253.2 253.2L225.7 763a31.325 31.325 0 0 0 0 44.4c6.1 6.1 14.2 9.2 22.2 9.2s16.1-3.1 22.2-9.2l253.2-253.2 253.2 253.2c6.1 6.1 14.2 9.2 22.2 9.2s16.1-3.1 22.2-9.2c12.3-12.3 12.3-32.1 0-44.4L567.7 509.8z"
-        fill={color}
-        p-id="4795"
-      ></path>
-    </svg>
-  );
-};
-
-export default CloseIcon;
