+++ app/routes/app.switcher/route.tsx
@@ -29,7 +29,6 @@ import SwitcherSettingCard from "./components/switcherSettingCard";
 const { Text, Title } = Typography;
 import defaultStyles from "../styles/defaultStyles.module.css";
 import useReport from "scripts/eventReport";
-import CloseIcon from "~/components/icon/closeIcon";
 interface EditData {
   shopName: string;
   includedFlag: boolean;
@@ -166,6 +165,8 @@ const Index = () => {
   const [currencySelector, setCurrencySelector] = useState(true);
   const [fontColor, setFontColor] = useState("#000000");
   const [backgroundColor, setBackgroundColor] = useState("#ffffff");
+  const [buttonColor, setButtonColor] = useState("#ffffff");
+  const [buttonBackgroundColor, setButtonBackgroundColor] = useState("#000000");
   const [optionBorderColor, setOptionBorderColor] = useState("#ccc");
   const [selectorPosition, setSelectorPosition] = useState("top_left");
   const [positionData, setPositionData] = useState<string>("0");
@@ -282,6 +283,8 @@ const Index = () => {
         setIsGeoLocationEnabled(res.ipOpen);
         setFontColor(res.fontColor);
         setBackgroundColor(res.backgroundColor);
+        setButtonColor(res.buttonColor);
+        setButtonBackgroundColor(res.buttonBackgroundColor);
         setOptionBorderColor(res.optionBorderColor);
         setSelectorPosition(res.selectorPosition);
         setPositionData(res.positionData);
@@ -296,6 +299,8 @@ const Index = () => {
         setIsGeoLocationEnabled(initData.ipOpen);
         setFontColor(initData.fontColor);
         setBackgroundColor(initData.backgroundColor);
+        setButtonColor(initData.buttonColor);
+        setButtonBackgroundColor(initData.buttonBackgroundColor);
         setOptionBorderColor(initData.optionBorderColor);
         setSelectorPosition(initData.selectorPosition);
         setPositionData(initData.positionData);
@@ -486,6 +491,12 @@ const Index = () => {
         case "backgroundColor":
           setBackgroundColor(value as string);
           break;
+        case "buttonColor":
+          setButtonColor(value as string);
+          break;
+        case "buttonBackgroundColor":
+          setButtonBackgroundColor(value as string);
+          break;
         case "optionBorderColor":
           setOptionBorderColor(value as string);
           break;
@@ -604,7 +615,6 @@ const Index = () => {
           language.selected = false;
         } else {
           language.selected = true;
-          setSelectedLanguage(language);
         }
       });
       setLocalization({ ...localization });
@@ -616,12 +626,10 @@ const Index = () => {
           currency.selected = false;
         } else {
           currency.selected = true;
-          setSelectedCurrency(currency);
         }
       });
       setLocalization({ ...localization });
     }
-    setIsSelectorOpen(false);
   };
 
   const handleIpOpenChange = (checked: boolean) => {
@@ -688,6 +696,8 @@ const Index = () => {
       setIsGeoLocationEnabled(originalData.ipOpen);
       setFontColor(originalData.fontColor);
       setBackgroundColor(originalData.backgroundColor);
+      setButtonColor(originalData.buttonColor);
+      setButtonBackgroundColor(originalData.buttonBackgroundColor);
       setOptionBorderColor(originalData.optionBorderColor);
       setSelectorPosition(originalData.selectorPosition);
       setPositionData(originalData.positionData);
@@ -939,6 +949,44 @@ const Index = () => {
                         justifyContent: "space-between",
                       }}
                     >
+                      <Text>{t("Button Color:")}</Text>
+                      <ColorPicker
+                        style={{ alignSelf: "flex-start" }}
+                        value={buttonColor}
+                        onChange={(e) =>
+                          handleEditData({ buttonColor: e.toHexString() })
+                        }
+                        showText
+                      />
+                    </div>
+                    <div
+                      style={{
+                        flex: 1,
+                        flexDirection: "column",
+                        display: "flex",
+                        justifyContent: "space-between",
+                      }}
+                    >
+                      <Text>{t("Button Background Color:")}</Text>
+                      <ColorPicker
+                        style={{ alignSelf: "flex-start" }}
+                        value={buttonBackgroundColor}
+                        onChange={(e) =>
+                          handleEditData({
+                            buttonBackgroundColor: e.toHexString(),
+                          })
+                        }
+                        showText
+                      />
+                    </div>
+                    <div
+                      style={{
+                        flex: 1,
+                        flexDirection: "column",
+                        display: "flex",
+                        justifyContent: "space-between",
+                      }}
+                    >
                       <Text>{t("Option Border Color:")}</Text>
                       <ColorPicker
                         style={{ alignSelf: "flex-start" }}
@@ -1047,10 +1095,11 @@ const Index = () => {
                   <div
                     id="ciwi-container"
                     style={{
-                      marginTop: "84px",
                       minWidth: "100px",
                       position: "absolute", // 改为绝对定位
                       top:
+                        selectorPosition === "top_left" ||
+                        selectorPosition === "top_right" ||
                         languageSelector === currencySelector
                           ? "-221px"
                           : "-171px", // 位于顶部
@@ -1076,32 +1125,16 @@ const Index = () => {
                         style={{
                           background: backgroundColor,
                           border: `1px solid ${optionBorderColor}`,
-                          padding: "10px",
+                          padding: "15px",
                           borderRadius: "5px",
-                          height:
-                            languageSelector === currencySelector
-                              ? "140px"
-                              : "90px",
-                          marginBottom: "1px",
+                          marginBottom: "5px",
                           width: "100%",
-                          display: "flex",
-                          flexDirection: "column",
-                          gap: "10px",
                         }}
                       >
-                        <div className={styles.close_button_wrapper}>
-                          <button
-                            onClick={handleCancelClick}
-                            className={styles.selector_box_close_button}
-                            id="selector-box-close-button"
-                          >
-                            <CloseIcon color={fontColor} />
-                          </button>
-                        </div>
                         <div
                           style={{
                             display: `${languageSelector || (!languageSelector && !currencySelector) ? "block" : "none"}`,
-                            gap: "10px",
+                            marginBottom: "10px",
                           }}
                         >
                           <div
@@ -1285,6 +1318,30 @@ const Index = () => {
                             </div>
                           </div>
                         </div>
+                        <div className={styles.button_wrapper}>
+                          <button
+                            id="switcher-confirm"
+                            className={styles.ciwi_switcher_confirm_button}
+                            style={{
+                              backgroundColor: buttonBackgroundColor,
+                              color: buttonColor,
+                            }}
+                            onClick={handleSelectorClick}
+                          >
+                            Confirm
+                          </button>
+                          <button
+                            id="switcher-cancel"
+                            className={styles.ciwi_switcher_confirm_button}
+                            style={{
+                              backgroundColor: buttonBackgroundColor,
+                              color: buttonColor,
+                            }}
+                            onClick={handleCancelClick}
+                          >
+                            Cancel
+                          </button>
+                        </div>
                       </div>
                     )}
                     <div
@@ -1294,8 +1351,8 @@ const Index = () => {
                         marginTop: isSelectorOpen
                           ? "0px"
                           : languageSelector === currencySelector
-                            ? "141px"
-                            : "91px",
+                            ? "225px"
+                            : "175px",
                         background: backgroundColor,
                         border: `1px solid ${optionBorderColor}`,
                         justifyContent: isIncludedFlag ? "" : "center",
@@ -1326,7 +1383,7 @@ const Index = () => {
                         className={styles.mainarrow_icon}
                         src="/arrow.svg"
                         alt="Arrow Icon"
-                        width="25px"
+                        width="15%"
                         height="25%"
                       />
                     </div>
