+++ extensions/ciwi-switcher/blocks/ciwi_I18n_Switcher.liquid
@@ -13,28 +13,12 @@
           class="selector_box"
           style="display: none;"
         >
-          <div class="close_button_wrapper">
-            <button class="selector_box_close_button" id="selector-box-close-button">
-              <svg
-                t="1759129501675"
-                class="icon"
-                viewBox="0 0 1024 1024"
-                version="1.1"
-                xmlns="http://www.w3.org/2000/svg"
-                p-id="4794"
-                width="16"
-                height="16"
-              >
-                <path d="M567.7 509.8l253.2-253.2c12.3-12.3 12.3-32.1 0-44.4a31.325 31.325 0 0 0-44.4 0L523.3 465.4 270.1 212.2a31.325 31.325 0 0 0-44.4 0 31.325 31.325 0 0 0 0 44.4l253.2 253.2L225.7 763a31.325 31.325 0 0 0 0 44.4c6.1 6.1 14.2 9.2 22.2 9.2s16.1-3.1 22.2-9.2l253.2-253.2 253.2 253.2c6.1 6.1 14.2 9.2 22.2 9.2s16.1-3.1 22.2-9.2c12.3-12.3 12.3-32.1 0-44.4L567.7 509.8z" fill="#2c2c2c" p-id="4795"></path>
-              </svg>
-            </button>
-          </div>
           <div
             id="language-switcher-container"
             style="display: none"
           >
             <div class="custom-selector" data-type="language">
-              <div class="language_selector_header" data-type="language">
+              <div class="selector-header" data-type="language">
                 <div class="selected-option" data-type="language">
                   <span
                     class="selected-text"
@@ -84,7 +68,7 @@
             style="display: none"
           >
             <div class="custom-selector" data-type="currency">
-              <div class="currency_selector_header" data-type="currency">
+              <div class="selector-header" data-type="currency">
                 <div class="selected-option" data-type="currency">
                   <span class="selected-text" data-type="currency"></span>
                 </div>
@@ -103,6 +87,10 @@
               </div>
             </div>
           </div>
+          <div class="button_wrapper">
+            <button id="switcher-confirm" class="ciwi_switcher_confirm_button">Confirm</button>
+            <button type="button" id="switcher-cancel" class="ciwi_switcher_cancel_button">Cancel</button>
+          </div>
         </div>
         <div
           id="main-box"
@@ -113,8 +101,8 @@
             id="main-language-flag"
             class="country-flag main_box_flag"
             alt=""
-            width="20px"
-            height="auto"
+            width="30px"
+            height="20px"
             src="data:image/gif;base64,R0lGODlhAQABAIAAAAAAAP///yH5BAEAAAAALAAAAAABAAEAAAIBRAA7"
             onerror="this.style.display='none'"
             style="margin-right: 8px; flex-shrink: 0;"
@@ -173,7 +161,7 @@
           </div>
           <img
             id="translate-float-btn-icon"
-            class="translate_float_btn_icon"
+            class="country-flag main_box_flag"
             alt=""
             width="28"
             height="20"
@@ -222,9 +210,10 @@
     {% endform %}
   </ciwiswitcher-form>
 
+  <script src="https://cdn.jsdelivr.net/npm/axios/dist/axios.min.js" async></script>
   <script src="{{ 'currency-format-config.js' | asset_url }}" async></script>
   <script src="{{ 'language-locale-data.js' | asset_url }}" async></script>
-  <script type="module" src="{{ 'ciwi-main.js' | asset_url }}"></script>
+  <script src="{{ 'switcher.js' | asset_url }}" async></script>
   <script>
     window.__JS_EXECUTED__ = true;
   </script>
